requirements.txt
livekit-agents~=1.5
livekit-api>=0.8.0
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-silero
python-dotenv>=1.0.1
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
